Backport upstream PR #13721 - Gui: Move more button to the end WB TabBar
Gui: Add our ToolBarArea enumeration

This refactors implementation of toolbars in menu / status bar a bit. It
introduces enum with all possible areas like it is in Qt that can be
later used to decide what to do based on toolbar placement.

Gui: Move more button to the end WB TabBar

This changes back placement of the "more" button of the WB TabBar to be
at the end, where it should be naturally placed. In order to ensure that
it is always visible the control was reworked to show this button always
after the tab bar widget which now is dynamically sized. This is
behavior that is well known from browsers.

This commit also ensures that active workbench is always visible in the
TabBar by adding additional temporary tab when necessary. This tab will
automatically dissapear when not needed.

Fixes: #13720
Fixes: #13630

Gui: Add proper tooltip to Workbench TabBar tab

Gui: Adjust sizing of Workbench TabBar when changing orientation

This should fix issues when toolbar containing Workbench TabBar suddenly
(or not) changes orientation. It also fixes size policies so toolbar
resizes properly and does not cause window to grow.

Fixes: #13286

Gui: Use RTL layout when Wb Tab Bar is placed in right corner

Right corner is placed to the right edge of screen, so its natural
growth occours on the left side. Basically it is Right to Left order and
so in that case the "end" is actually on left and so TabBar should grow
in that direction.

Unfortunately it is not possible to simply use RTL Qt feature to handle
that case as it would result in reverse order of workbenches (people will still
read it in LTR order) and icons on the right which is not wanted. That's
custom support is introduced.

Gui: Fix wrong orientation of workbench tab bar after start

This replaces old mechanism that was based on storing tab bar
orientation in user settings with one that delays initialization by half
of a second to ensure that toolbar is placed where in right place.

Gui: Reuse QActions for workbench activation

This fixes segfault that can occour due to keeping reference to QAction
that is supposed to change workbench.