Gui: Handle property types better in VarSet dialog
In particular, this commit makes a distinction between types that should
not have an editor (editing values can be done in the property view) and
types that are not supported, which are types that throw exceptions.
These types are more for internal use, for example as base classes.
Examples are App::Property and App::PropertyLinkBase.